Updated: UI
I have created the healthbar method from the tutorial. I don't really like it, but I will see if the rest of the tut gives the rest of the answers.

Along with the healthbar i have also added an info panel which appears alongside a button when it is hovered over.