Update: Selecction Method
The tutorial I am currently following is working through re-vamping my selection method, so that a view of what units are selected, even if they are off screen.

So far, I have worked out getting an object to represent how many of each unit type i have selected, however when selecting or deselecting units I am getting a NULL EXEMPTION ERROR, this does not seem to stop the program from running. I am now going to complete the rest of this tutorial and hope that it fixes the error, because i have been unable to find the cause.